project/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.3.9
channels==2.2.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
daphne==2.3.0
Django==2.2
django-admin-autocomplete-filter==0.2
django-autocomplete-light==3.3.4
django-celery==3.2.2
django-extensions==2.1.6
django-filter==2.1.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.9.2
djoser==1.5.1
forex-python==1.5
future==0.17.1
googlemaps==3.0.2
gunicorn==19.9.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10.1
kombu==3.0.37
line-bot-sdk==1.8.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
pika==1.0.1
Pillow==6.0.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
Pygments==2.3.1
PyHamcrest==1.9.0
pyTelegramBotAPI==3.6.6
pytz==2019.1
requests==2.21.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.24.2
vine==1.3.0
zope.interface==4.6.0
